extract investments feature and move csv import
