+++ frontend/src/App.tsx
@@ -2,25 +2,19 @@ import {
   startTransition,
   useEffect,
   useMemo,
-  useRef,
   useState,
-  type FormEvent,
 } from 'react'
 import {
   ApiClientError,
   apiClient,
-  type AccountDto,
-  type AccountType,
   type CreatePersonalFinanceCardRequest,
   type CreatePersonalFinanceTransferRequest,
   type CreateAccountRequest,
   type CreateTransactionRequest,
   type CurrencyTotalsDto,
   type ImportTransactionsCsvResponse,
-  type MoneyDto,
   type PersonalFinanceCardDto,
   type PersonalFinanceSnapshotDto,
-  type TransactionDirection,
   type TransactionDto,
   type UpdateIncomePlanRequest,
   type UpdateAccountRequest,
@@ -31,12 +25,25 @@ import {
   type UpdateTransactionRequest,
 } from './api'
 import {
+  InvestmentsView,
+  formatIsoDateRu,
+  normalizeTransactionMemo,
+  todayIsoDate,
+  MetricCard,
+  StatusCard,
+  type AccountWithBalance,
+  type CreateAccountStatus,
+  type CreateTransactionStatus,
+  type CsvImportStatus,
+  type LoadStatus,
+  type TransactionDirectionFilter,
+} from './features/investments'
+import {
   PersonalFinanceView,
   type PersonalFinanceCardListItem,
   type PersonalFinanceTab,
 } from './features/personal-finance'
 import { AppShell, type ViewTab } from './features/app-shell'
-import { formatMoneyInput, normalizeMoneyInput } from './money-input'
 
 interface DashboardData {
   asOf: string
@@ -45,16 +52,6 @@ interface DashboardData {
   monthlySavings: CurrencyTotalsDto
 }
 
-interface AccountWithBalance extends AccountDto {
-  balance: MoneyDto
-}
-
-type LoadStatus = 'idle' | 'loading' | 'ready' | 'error'
-type TransactionDirectionFilter = 'ALL' | 'INFLOW' | 'OUTFLOW'
-type CreateAccountStatus = 'idle' | 'submitting' | 'error'
-type CreateTransactionStatus = 'idle' | 'submitting' | 'error'
-type CsvImportStatus = 'idle' | 'submitting' | 'success' | 'error'
-type AccountStatus = AccountDto['status']
 interface NavigationState {
   tab: ViewTab
   accountId: string | null
@@ -63,19 +60,6 @@ interface NavigationState {
   financeCardId: string | null
 }
 
-const ACCOUNT_TYPE_OPTIONS: AccountType[] = [
-  'CASH',
-  'DEPOSIT',
-  'FUND',
-  'IIS',
-  'BROKERAGE',
-]
-const ACCOUNT_CURRENCY_OPTIONS: string[] = getAccountCurrencyOptions()
-const DEFAULT_ACCOUNT_CURRENCY =
-  ACCOUNT_CURRENCY_OPTIONS.includes('USD') &&
-  ACCOUNT_CURRENCY_OPTIONS.length > 0
-    ? 'USD'
-    : (ACCOUNT_CURRENCY_OPTIONS[0] ?? 'USD')
 const DEFAULT_NAVIGATION_STATE: NavigationState = {
   tab: 'dashboard',
   accountId: null,
@@ -855,7 +839,7 @@ function App() {
           onDeleteCard={handleDeletePersonalFinanceCard}
         />
       ) : (
-        <AccountsView
+        <InvestmentsView
           status={accountsStatus}
           accounts={accounts}
           selectedAccountId={selectedAccountId}
@@ -955,1308 +939,6 @@ function DashboardView({
       </div>
     </div>
   )
-}
-
-interface AccountsViewProps {
-  status: LoadStatus
-  accounts: AccountWithBalance[]
-  selectedAccountId: string | null
-  selectedAccount: AccountWithBalance | null
-  transactionsStatus: LoadStatus
-  transactionsErrorMessage: string | null
-  filteredTransactions: TransactionDto[]
-  totalTransactionsCount: number
-  directionFilter: TransactionDirectionFilter
-  memoFilter: string
-  onDirectionFilterChange: (value: TransactionDirectionFilter) => void
-  onMemoFilterChange: (value: string) => void
-  onSelectAccount: (accountId: string) => void
-  createAccountStatus: CreateAccountStatus
-  createAccountErrorMessage: string | null
-  onCreateAccount: (request: CreateAccountRequest) => Promise<boolean>
-  onUpdateAccount: (
-    accountId: string,
-    request: UpdateAccountRequest,
-  ) => Promise<void>
-  onDeleteAccount: (accountId: string) => Promise<void>
-  createTransactionStatus: CreateTransactionStatus
-  createTransactionErrorMessage: string | null
-  onCreateTransaction: (
-    accountId: string,
-    request: CreateTransactionRequest,
-  ) => Promise<boolean>
-  onUpdateTransaction: (
-    accountId: string,
-    transactionId: string,
-    request: UpdateTransactionRequest,
-  ) => Promise<void>
-  onDeleteTransaction: (
-    accountId: string,
-    transactionId: string,
-  ) => Promise<void>
-  csvImportStatus: CsvImportStatus
-  csvImportErrorMessage: string | null
-  csvImportResult: ImportTransactionsCsvResponse | null
-  onImportTransactionsCsv: (accountId: string, file: File) => Promise<boolean>
-  errorMessage: string | null
-  onRetryAccounts: () => void
-  onRetryTransactions: () => void
-}
-
-function AccountsView({
-  status,
-  accounts,
-  selectedAccountId,
-  selectedAccount,
-  transactionsStatus,
-  transactionsErrorMessage,
-  filteredTransactions,
-  totalTransactionsCount,
-  directionFilter,
-  memoFilter,
-  onDirectionFilterChange,
-  onMemoFilterChange,
-  onSelectAccount,
-  createAccountStatus,
-  createAccountErrorMessage,
-  onCreateAccount,
-  onUpdateAccount,
-  onDeleteAccount,
-  createTransactionStatus,
-  createTransactionErrorMessage,
-  onCreateTransaction,
-  onUpdateTransaction,
-  onDeleteTransaction,
-  csvImportStatus,
-  csvImportErrorMessage,
-  csvImportResult,
-  onImportTransactionsCsv,
-  errorMessage,
-  onRetryAccounts,
-  onRetryTransactions,
-}: AccountsViewProps) {
-  const [newAccountName, setNewAccountName] = useState<string>('')
-  const [newAccountCurrency, setNewAccountCurrency] = useState<string>(
-    DEFAULT_ACCOUNT_CURRENCY,
-  )
-  const [newAccountType, setNewAccountType] = useState<AccountType>('CASH')
-  const [editingAccountId, setEditingAccountId] = useState<string | null>(null)
-  const [editingAccountName, setEditingAccountName] = useState<string>('')
-  const [editingAccountType, setEditingAccountType] =
-    useState<AccountType>('CASH')
-  const [updateAccountStatus, setUpdateAccountStatus] =
-    useState<CreateAccountStatus>('idle')
-  const [updateAccountErrorMessage, setUpdateAccountErrorMessage] = useState<
-    string | null
-  >(null)
-  const [deleteAccountStatus, setDeleteAccountStatus] =
-    useState<CreateAccountStatus>('idle')
-  const [deleteAccountErrorMessage, setDeleteAccountErrorMessage] = useState<
-    string | null
-  >(null)
-  const [deleteAccountErrorAccountId, setDeleteAccountErrorAccountId] =
-    useState<string | null>(null)
-  const [newTransactionDate, setNewTransactionDate] =
-    useState<string>(todayIsoDate())
-  const [newTransactionDirection, setNewTransactionDirection] =
-    useState<TransactionDirection>('OUTFLOW')
-  const [newTransactionAmount, setNewTransactionAmount] = useState<string>('')
-  const [newTransactionMemo, setNewTransactionMemo] = useState<string>('')
-  const [editingTransactionAccountId, setEditingTransactionAccountId] =
-    useState<string | null>(null)
-  const [editingTransactionId, setEditingTransactionId] = useState<
-    string | null
-  >(null)
-  const [editingTransactionDate, setEditingTransactionDate] =
-    useState<string>(todayIsoDate())
-  const [editingTransactionDirection, setEditingTransactionDirection] =
-    useState<TransactionDirection>('OUTFLOW')
-  const [editingTransactionAmount, setEditingTransactionAmount] =
-    useState<string>('')
-  const [editingTransactionMemo, setEditingTransactionMemo] =
-    useState<string>('')
-  const [updateTransactionStatus, setUpdateTransactionStatus] =
-    useState<CreateTransactionStatus>('idle')
-  const [updateTransactionErrorMessage, setUpdateTransactionErrorMessage] =
-    useState<string | null>(null)
-  const [deletingTransactionId, setDeletingTransactionId] = useState<
-    string | null
-  >(null)
-  const [deleteTransactionErrorMessage, setDeleteTransactionErrorMessage] =
-    useState<string | null>(null)
-  const [csvFile, setCsvFile] = useState<File | null>(null)
-  const csvFileInputRef = useRef<HTMLInputElement | null>(null)
-
-  const resetEditingAccount = (): void => {
-    setEditingAccountId(null)
-    setEditingAccountName('')
-    setEditingAccountType('CASH')
-    setUpdateAccountStatus('idle')
-    setUpdateAccountErrorMessage(null)
-  }
-
-  const resetDeleteAccountState = (): void => {
-    setDeleteAccountStatus('idle')
-    setDeleteAccountErrorMessage(null)
-    setDeleteAccountErrorAccountId(null)
-  }
-
-  const resetEditingTransaction = (): void => {
-    setEditingTransactionAccountId(null)
-    setEditingTransactionId(null)
-    setEditingTransactionDate(todayIsoDate())
-    setEditingTransactionDirection('OUTFLOW')
-    setEditingTransactionAmount('')
-    setEditingTransactionMemo('')
-    setUpdateTransactionStatus('idle')
-    setUpdateTransactionErrorMessage(null)
-  }
-
-  const resetDeleteTransactionState = (): void => {
-    setDeletingTransactionId(null)
-    setDeleteTransactionErrorMessage(null)
-  }
-
-  if (status === 'loading' || status === 'idle') {
-    return (
-      <StatusCard tone="neutral" message="Загружаем инвестиции и балансы..." />
-    )
-  }
-
-  if (status === 'error') {
-    return (
-      <StatusCard
-        tone="warning"
-        message={errorMessage ?? 'Не удалось загрузить инвестиции.'}
-        actionLabel="Повторить"
-        onAction={onRetryAccounts}
-      />
-    )
-  }
-
-  const isCurrencyValid = ACCOUNT_CURRENCY_OPTIONS.includes(newAccountCurrency)
-  const canCreate =
-    newAccountName.trim().length > 0 &&
-    isCurrencyValid &&
-    createAccountStatus !== 'submitting'
-  const isEditingSelectedAccount =
-    selectedAccount !== null && editingAccountId === selectedAccount.id
-  const canUpdateAccount =
-    isEditingSelectedAccount &&
-    editingAccountName.trim().length > 0 &&
-    updateAccountStatus !== 'submitting'
-  const isDeleteAvailabilityKnown = transactionsStatus === 'ready'
-  const hasLoadedTransactions =
-    isDeleteAvailabilityKnown && totalTransactionsCount > 0
-  const canDeleteAccount =
-    selectedAccount !== null &&
-    !isEditingSelectedAccount &&
-    isDeleteAvailabilityKnown &&
-    totalTransactionsCount === 0 &&
-    deleteAccountStatus !== 'submitting'
-
-  const transactionDateCandidate = newTransactionDate.trim()
-  const transactionAmountCandidate = normalizeMoneyInput(newTransactionAmount)
-  const transactionMemoCandidate = newTransactionMemo.trim()
-  const isTransactionDateValid = isValidIsoDateValue(transactionDateCandidate)
-  const isTransactionAmountValid = isValidPositiveAmountValue(
-    transactionAmountCandidate,
-  )
-  const canCreateTransaction =
-    selectedAccount !== null &&
-    isTransactionDateValid &&
-    isTransactionAmountValid &&
-    createTransactionStatus !== 'submitting'
-  const editingTransactionDateCandidate = editingTransactionDate.trim()
-  const editingTransactionAmountCandidate = normalizeMoneyInput(
-    editingTransactionAmount,
-  )
-  const editingTransactionMemoCandidate = editingTransactionMemo.trim()
-  const isEditingTransactionDateValid = isValidIsoDateValue(
-    editingTransactionDateCandidate,
-  )
-  const isEditingTransactionAmountValid = isValidPositiveAmountValue(
-    editingTransactionAmountCandidate,
-  )
-  const activeEditingTransactionId =
-    editingTransactionAccountId === selectedAccountId
-      ? editingTransactionId
-      : null
-  const canUpdateTransaction =
-    selectedAccount !== null &&
-    activeEditingTransactionId !== null &&
-    isEditingTransactionDateValid &&
-    isEditingTransactionAmountValid &&
-    updateTransactionStatus !== 'submitting'
-  const canImportCsv =
-    selectedAccount !== null &&
-    csvFile !== null &&
-    csvImportStatus !== 'submitting'
-  const isDeleteTransactionSubmitting =
-    deletingTransactionId !== null && deleteTransactionErrorMessage === null
-
-  const handleCreateAccountSubmit = async (
-    event: FormEvent<HTMLFormElement>,
-  ): Promise<void> => {
-    event.preventDefault()
-
-    if (!canCreate) {
-      return
-    }
-
-    const created = await onCreateAccount({
-      name: newAccountName.trim(),
-      currency: newAccountCurrency,
-      type: newAccountType,
-    })
-
-    if (created) {
-      setNewAccountName('')
-    }
-  }
-
-  const handleStartEditingAccount = (): void => {
-    if (!selectedAccount) {
-      return
-    }
-
-    resetDeleteAccountState()
-    setEditingAccountId(selectedAccount.id)
-    setEditingAccountName(selectedAccount.name)
-    setEditingAccountType(selectedAccount.type)
-    setUpdateAccountStatus('idle')
-    setUpdateAccountErrorMessage(null)
-  }
-
-  const handleUpdateAccountSubmit = async (
-    event: FormEvent<HTMLFormElement>,
-  ): Promise<void> => {
-    event.preventDefault()
-
-    if (!selectedAccount || !isEditingSelectedAccount || !canUpdateAccount) {
-      return
-    }
-
-    resetDeleteAccountState()
-    setUpdateAccountStatus('submitting')
-    setUpdateAccountErrorMessage(null)
-
-    try {
-      await onUpdateAccount(selectedAccount.id, {
-        name: editingAccountName.trim(),
-        type: editingAccountType,
-      })
-      resetEditingAccount()
-    } catch (error) {
-      setUpdateAccountStatus('error')
-      setUpdateAccountErrorMessage(toErrorMessage(error))
-    }
-  }
-
-  const handleDeleteAccountClick = async (): Promise<void> => {
-    if (!selectedAccount || !canDeleteAccount) {
-      return
-    }
-
-    if (
-      !window.confirm(
-        `Удалить счет «${selectedAccount.name}»? Это действие необратимо.`,
-      )
-    ) {
-      return
-    }
-
-    setDeleteAccountStatus('submitting')
-    setDeleteAccountErrorMessage(null)
-
-    try {
-      await onDeleteAccount(selectedAccount.id)
-      resetEditingAccount()
-      resetEditingTransaction()
-      resetDeleteTransactionState()
-      resetDeleteAccountState()
-    } catch (error) {
-      setDeleteAccountStatus('error')
-      setDeleteAccountErrorMessage(toErrorMessage(error))
-      setDeleteAccountErrorAccountId(selectedAccount.id)
-    }
-  }
-
-  const handleCreateTransactionSubmit = async (
-    event: FormEvent<HTMLFormElement>,
-  ): Promise<void> => {
-    event.preventDefault()
-
-    if (!selectedAccount || !canCreateTransaction) {
-      return
-    }
-
-    resetDeleteAccountState()
-    const created = await onCreateTransaction(selectedAccount.id, {
-      occurredOn: transactionDateCandidate,
-      direction: newTransactionDirection,
-      amount: transactionAmountCandidate,
-      memo: transactionMemoCandidate,
-    })
-
-    if (created) {
-      setNewTransactionAmount('')
-      setNewTransactionMemo('')
-    }
-  }
-
-  const handleImportCsvSubmit = async (
-    event: FormEvent<HTMLFormElement>,
-  ): Promise<void> => {
-    event.preventDefault()
-
-    if (!selectedAccount || !csvFile || !canImportCsv) {
-      return
-    }
-
-    resetDeleteAccountState()
-    const imported = await onImportTransactionsCsv(selectedAccount.id, csvFile)
-    if (!imported) {
-      return
-    }
-
-    setCsvFile(null)
-    if (csvFileInputRef.current) {
-      csvFileInputRef.current.value = ''
-    }
-  }
-
-  const handleStartEditingTransaction = (transaction: TransactionDto): void => {
-    resetDeleteTransactionState()
-    setEditingTransactionAccountId(selectedAccountId)
-    setEditingTransactionId(transaction.id)
-    setEditingTransactionDate(transaction.occurredOn)
-    setEditingTransactionDirection(transaction.direction)
-    setEditingTransactionAmount(formatMoneyInput(transaction.amount))
-    setEditingTransactionMemo(normalizeTransactionMemo(transaction.memo))
-    setUpdateTransactionStatus('idle')
-    setUpdateTransactionErrorMessage(null)
-  }
-
-  const handleUpdateTransactionSubmit = async (
-    event: FormEvent<HTMLFormElement>,
-  ): Promise<void> => {
-    event.preventDefault()
-
-    if (
-      !selectedAccount ||
-      !activeEditingTransactionId ||
-      !canUpdateTransaction
-    ) {
-      return
-    }
-
-    resetDeleteAccountState()
-    setUpdateTransactionStatus('submitting')
-    setUpdateTransactionErrorMessage(null)
-
-    try {
-      await onUpdateTransaction(
-        selectedAccount.id,
-        activeEditingTransactionId,
-        {
-          occurredOn: editingTransactionDateCandidate,
-          direction: editingTransactionDirection,
-          amount: editingTransactionAmountCandidate,
-          memo: editingTransactionMemoCandidate,
-        },
-      )
-      resetEditingTransaction()
-    } catch (error) {
-      setUpdateTransactionStatus('error')
-      setUpdateTransactionErrorMessage(toErrorMessage(error))
-    }
-  }
-
-  const handleDeleteTransactionClick = async (
-    transaction: TransactionDto,
-  ): Promise<void> => {
-    if (!selectedAccount || isDeleteTransactionSubmitting) {
-      return
-    }
-
-    resetDeleteAccountState()
-    setDeletingTransactionId(transaction.id)
-    setDeleteTransactionErrorMessage(null)
-
-    try {
-      await onDeleteTransaction(selectedAccount.id, transaction.id)
-      if (activeEditingTransactionId === transaction.id) {
-        resetEditingTransaction()
-      }
-      resetDeleteTransactionState()
-    } catch (error) {
-      setDeletingTransactionId(transaction.id)
-      setDeleteTransactionErrorMessage(toErrorMessage(error))
-    }
-  }
-
-  const handleSelectAccountClick = (accountId: string): void => {
-    onSelectAccount(accountId)
-    resetEditingAccount()
-    resetDeleteAccountState()
-    resetEditingTransaction()
-    resetDeleteTransactionState()
-    setNewTransactionDate(todayIsoDate())
-    setNewTransactionDirection('OUTFLOW')
-    setNewTransactionAmount('')
-    setNewTransactionMemo('')
-    setCsvFile(null)
-    if (csvFileInputRef.current) {
-      csvFileInputRef.current.value = ''
-    }
-  }
-
-  return (
-    <div className="grid gap-4 lg:grid-cols-[minmax(0,0.9fr)_minmax(0,1.3fr)]">
-      <article className="rounded-xl border border-slate-200 bg-slate-50 p-4">
-        <h2 className="text-sm font-semibold text-slate-900">
-          Список инвестиций
-        </h2>
-        <p className="mt-1 text-xs text-slate-500">
-          Выберите инструмент, чтобы посмотреть транзакции.
-        </p>
-
-        <form
-          className="mt-4 rounded-lg border border-slate-200 bg-white p-3"
-          onSubmit={(event) => {
-            void handleCreateAccountSubmit(event)
-          }}
-        >
-          <p className="text-xs font-medium uppercase tracking-wide text-slate-500">
-            Добавить инструмент
-          </p>
-
-          <div className="mt-3 space-y-2">
-            <input
-              type="text"
-              value={newAccountName}
-              onChange={(event) => setNewAccountName(event.target.value)}
-              placeholder="Название инструмента"
-              className="block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-            />
-
-            <div className="grid gap-2 sm:grid-cols-2">
-              <select
-                value={newAccountCurrency}
-                onChange={(event) => setNewAccountCurrency(event.target.value)}
-                className="block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-              >
-                {ACCOUNT_CURRENCY_OPTIONS.map((currencyCode) => (
-                  <option key={currencyCode} value={currencyCode}>
-                    {currencyCode}
-                  </option>
-                ))}
-              </select>
-
-              <select
-                value={newAccountType}
-                onChange={(event) =>
-                  setNewAccountType(event.target.value as AccountType)
-                }
-                className="block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-              >
-                {ACCOUNT_TYPE_OPTIONS.map((type) => (
-                  <option key={type} value={type}>
-                    {toAccountTypeLabel(type)}
-                  </option>
-                ))}
-              </select>
-            </div>
-          </div>
-
-          {!isCurrencyValid ? (
-            <p className="mt-2 text-xs text-amber-700">
-              Выберите валюту из списка.
-            </p>
-          ) : null}
-
-          {createAccountStatus === 'error' && createAccountErrorMessage ? (
-            <p className="mt-2 text-xs text-amber-700">
-              {createAccountErrorMessage}
-            </p>
-          ) : null}
-
-          <button
-            type="submit"
-            disabled={!canCreate}
-            className="mt-3 rounded-md border border-slate-300 bg-slate-100 px-3 py-1.5 text-xs font-medium text-slate-800 disabled:cursor-not-allowed disabled:opacity-50"
-          >
-            {createAccountStatus === 'submitting' ? 'Добавляем...' : 'Добавить'}
-          </button>
-        </form>
-
-        {accounts.length === 0 ? (
-          <div className="mt-4">
-            <StatusCard
-              tone="neutral"
-              message="Пока нет инвестиций. Добавьте первый инструмент."
-            />
-          </div>
-        ) : (
-          <ul className="mt-4 space-y-2">
-            {accounts.map((account) => (
-              <li key={account.id}>
-                <button
-                  type="button"
-                  onClick={() => handleSelectAccountClick(account.id)}
-                  className={`w-full rounded-lg border px-3 py-3 text-left ${
-                    selectedAccountId === account.id
-                      ? 'border-slate-300 bg-white shadow-sm'
-                      : 'border-slate-200 bg-transparent'
-                  }`}
-                >
-                  <div className="flex items-center justify-between gap-3">
-                    <div>
-                      <p className="text-sm font-semibold text-slate-900">
-                        {account.name}
-                      </p>
-                      <p className="mt-1 text-xs uppercase tracking-wide text-slate-500">
-                        {toAccountTypeLabel(account.type)} ·{' '}
-                        {toAccountStatusLabel(account.status)}
-                      </p>
-                    </div>
-                    <div className="text-right">
-                      <p className="text-xs uppercase tracking-wide text-slate-500">
-                        {account.balance.currency}
-                      </p>
-                      <p className="mt-1 text-sm font-semibold text-slate-900">
-                        {formatAmount(account.balance.amount)}
-                      </p>
-                    </div>
-                  </div>
-                </button>
-              </li>
-            ))}
-          </ul>
-        )}
-      </article>
-
-      <article className="rounded-xl border border-slate-200 bg-slate-50 p-4">
-        {!selectedAccount ? (
-          <StatusCard
-            tone="neutral"
-            message="Выберите инструмент, чтобы посмотреть транзакции."
-          />
-        ) : (
-          <>
-            <div className="flex flex-wrap items-center justify-between gap-2">
-              <div>
-                <h2 className="text-sm font-semibold text-slate-900">
-                  {selectedAccount.name}
-                </h2>
-                <p className="mt-1 text-xs uppercase tracking-wide text-slate-500">
-                  Транзакции · {selectedAccount.balance.currency}
-                </p>
-              </div>
-              {!isEditingSelectedAccount ? (
-                <div className="flex flex-wrap gap-2">
-                  <button
-                    type="button"
-                    onClick={handleStartEditingAccount}
-                    disabled={deleteAccountStatus === 'submitting'}
-                    className="rounded-md border border-slate-300 bg-slate-100 px-3 py-1.5 text-xs font-medium text-slate-800 disabled:cursor-not-allowed disabled:opacity-50"
-                  >
-                    Редактировать счет
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() => {
-                      void handleDeleteAccountClick()
-                    }}
-                    disabled={!canDeleteAccount}
-                    className="rounded-md border border-rose-200 bg-rose-50 px-3 py-1.5 text-xs font-medium text-rose-700 disabled:cursor-not-allowed disabled:opacity-50"
-                  >
-                    {deleteAccountStatus === 'submitting'
-                      ? 'Удаляем счет...'
-                      : 'Удалить счет'}
-                  </button>
-                </div>
-              ) : null}
-            </div>
-
-            {isEditingSelectedAccount ? (
-              <form
-                className="mt-4 rounded-lg border border-slate-200 bg-white p-3"
-                onSubmit={(event) => {
-                  void handleUpdateAccountSubmit(event)
-                }}
-              >
-                <p className="text-xs font-medium uppercase tracking-wide text-slate-500">
-                  Редактировать счет
-                </p>
-
-                <div className="mt-3 grid gap-2 sm:grid-cols-2">
-                  <label className="text-xs text-slate-600">
-                    Название
-                    <input
-                      type="text"
-                      value={editingAccountName}
-                      onChange={(event) =>
-                        setEditingAccountName(event.target.value)
-                      }
-                      className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                    />
-                  </label>
-
-                  <label className="text-xs text-slate-600">
-                    Тип
-                    <select
-                      value={editingAccountType}
-                      onChange={(event) =>
-                        setEditingAccountType(event.target.value as AccountType)
-                      }
-                      className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                    >
-                      {ACCOUNT_TYPE_OPTIONS.map((type) => (
-                        <option key={type} value={type}>
-                          {toAccountTypeLabel(type)}
-                        </option>
-                      ))}
-                    </select>
-                  </label>
-                </div>
-
-                <p className="mt-2 text-xs text-slate-500">
-                  Валюта счета фиксируется при создании и не редактируется:{' '}
-                  {selectedAccount.currency}.
-                </p>
-
-                {updateAccountStatus === 'error' &&
-                updateAccountErrorMessage ? (
-                  <p className="mt-2 text-xs text-amber-700">
-                    {updateAccountErrorMessage}
-                  </p>
-                ) : null}
-
-                <div className="mt-3 flex flex-wrap gap-2">
-                  <button
-                    type="submit"
-                    disabled={!canUpdateAccount}
-                    className="rounded-md border border-slate-300 bg-slate-100 px-3 py-1.5 text-xs font-medium text-slate-800 disabled:cursor-not-allowed disabled:opacity-50"
-                  >
-                    {updateAccountStatus === 'submitting'
-                      ? 'Сохраняем...'
-                      : 'Сохранить'}
-                  </button>
-                  <button
-                    type="button"
-                    onClick={resetEditingAccount}
-                    className="rounded-md border border-slate-200 px-3 py-1.5 text-xs font-medium text-slate-600"
-                  >
-                    Отмена
-                  </button>
-                </div>
-              </form>
-            ) : (
-              <p className="mt-2 text-xs text-slate-500">
-                Валюта счета фиксируется при создании и не редактируется.
-              </p>
-            )}
-
-            {!isEditingSelectedAccount && !isDeleteAvailabilityKnown ? (
-              <p className="mt-2 text-xs text-slate-500">
-                Удаление станет доступно после загрузки транзакций счета.
-              </p>
-            ) : null}
-
-            {!isEditingSelectedAccount && hasLoadedTransactions ? (
-              <p className="mt-2 text-xs text-slate-500">
-                Удаление доступно только для пустого счета без транзакций.
-              </p>
-            ) : null}
-
-            {!isEditingSelectedAccount &&
-            deleteAccountStatus === 'error' &&
-            deleteAccountErrorMessage &&
-            deleteAccountErrorAccountId === selectedAccount.id ? (
-              <p className="mt-2 text-xs text-amber-700">
-                {deleteAccountErrorMessage}
-              </p>
-            ) : null}
-
-            <form
-              className="mt-4 rounded-lg border border-slate-200 bg-white p-3"
-              onSubmit={(event) => {
-                void handleCreateTransactionSubmit(event)
-              }}
-            >
-              <p className="text-xs font-medium uppercase tracking-wide text-slate-500">
-                Добавить транзакцию
-              </p>
-
-              <div className="mt-3 grid gap-2 sm:grid-cols-2">
-                <label className="text-xs text-slate-600">
-                  Дата
-                  <input
-                    type="date"
-                    value={newTransactionDate}
-                    onChange={(event) =>
-                      setNewTransactionDate(event.target.value)
-                    }
-                    className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                  />
-                </label>
-
-                <label className="text-xs text-slate-600">
-                  Направление
-                  <select
-                    value={newTransactionDirection}
-                    onChange={(event) =>
-                      setNewTransactionDirection(
-                        event.target.value as TransactionDirection,
-                      )
-                    }
-                    className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                  >
-                    <option value="OUTFLOW">
-                      {toDirectionSelectLabel('OUTFLOW')}
-                    </option>
-                    <option value="INFLOW">
-                      {toDirectionSelectLabel('INFLOW')}
-                    </option>
-                  </select>
-                </label>
-              </div>
-
-              <div className="mt-2 grid gap-2 sm:grid-cols-[minmax(0,0.6fr)_minmax(0,1fr)]">
-                <label className="text-xs text-slate-600">
-                  Сумма
-                  <input
-                    type="text"
-                    inputMode="decimal"
-                    value={newTransactionAmount}
-                    onChange={(event) =>
-                      setNewTransactionAmount(
-                        formatMoneyInput(event.target.value),
-                      )
-                    }
-                    placeholder="0,00"
-                    className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                  />
-                </label>
-
-                <label className="text-xs text-slate-600">
-                  Описание
-                  <input
-                    type="text"
-                    value={newTransactionMemo}
-                    onChange={(event) =>
-                      setNewTransactionMemo(event.target.value)
-                    }
-                    placeholder="Продукты"
-                    className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                  />
-                </label>
-              </div>
-
-              {!isTransactionAmountValid &&
-              newTransactionAmount.trim().length > 0 ? (
-                <p className="mt-2 text-xs text-amber-700">
-                  Сумма должна быть положительной и содержать не более 2 знаков
-                  после запятой.
-                </p>
-              ) : null}
-
-              {createTransactionStatus === 'error' &&
-              createTransactionErrorMessage ? (
-                <p className="mt-2 text-xs text-amber-700">
-                  {createTransactionErrorMessage}
-                </p>
-              ) : null}
-
-              <button
-                type="submit"
-                disabled={!canCreateTransaction}
-                className="mt-3 rounded-md border border-slate-300 bg-slate-100 px-3 py-1.5 text-xs font-medium text-slate-800 disabled:cursor-not-allowed disabled:opacity-50"
-              >
-                {createTransactionStatus === 'submitting'
-                  ? 'Добавляем...'
-                  : 'Добавить транзакцию'}
-              </button>
-            </form>
-
-            <form
-              className="mt-4 rounded-lg border border-slate-200 bg-white p-3"
-              onSubmit={(event) => {
-                void handleImportCsvSubmit(event)
-              }}
-            >
-              <p className="text-xs font-medium uppercase tracking-wide text-slate-500">
-                Импорт из CSV
-              </p>
-              <p className="mt-1 text-xs text-slate-500">
-                Формат заголовка:{' '}
-                <code>occurred_on,direction,amount,currency,memo</code>
-              </p>
-
-              <label className="mt-3 block text-xs text-slate-600">
-                Файл
-                <input
-                  ref={csvFileInputRef}
-                  type="file"
-                  accept=".csv,text/csv"
-                  onChange={(event) =>
-                    setCsvFile(event.target.files?.[0] ?? null)
-                  }
-                  className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800 file:mr-3 file:rounded-md file:border-0 file:bg-slate-100 file:px-2 file:py-1 file:text-xs file:font-medium file:text-slate-700"
-                />
-              </label>
-
-              {csvFile ? (
-                <p className="mt-2 text-xs text-slate-500">
-                  Выбран файл:{' '}
-                  <span className="font-medium text-slate-700">
-                    {csvFile.name}
-                  </span>
-                </p>
-              ) : null}
-
-              {csvImportStatus === 'error' && csvImportErrorMessage ? (
-                <p className="mt-2 text-xs text-amber-700">
-                  {csvImportErrorMessage}
-                </p>
-              ) : null}
-
-              {csvImportStatus === 'success' && csvImportResult ? (
-                <p className="mt-2 text-xs text-emerald-700">
-                  Импорт завершен: получено {csvImportResult.receivedRows},
-                  добавлено {csvImportResult.importedCount}, пропущено
-                  дубликатов {csvImportResult.skippedDuplicates}.
-                </p>
-              ) : null}
-
-              <button
-                type="submit"
-                disabled={!canImportCsv}
-                className="mt-3 rounded-md border border-slate-300 bg-slate-100 px-3 py-1.5 text-xs font-medium text-slate-800 disabled:cursor-not-allowed disabled:opacity-50"
-              >
-                {csvImportStatus === 'submitting'
-                  ? 'Импортируем...'
-                  : 'Импортировать CSV'}
-              </button>
-            </form>
-
-            <div className="mt-4 grid gap-3 sm:grid-cols-2">
-              <label className="text-xs text-slate-600">
-                Направление
-                <select
-                  value={directionFilter}
-                  onChange={(event) =>
-                    onDirectionFilterChange(
-                      event.target.value as TransactionDirectionFilter,
-                    )
-                  }
-                  className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                >
-                  <option value="ALL">Все</option>
-                  <option value="INFLOW">Доходы</option>
-                  <option value="OUTFLOW">Расходы</option>
-                </select>
-              </label>
-
-              <label className="text-xs text-slate-600">
-                Фильтр по описанию
-                <input
-                  type="text"
-                  value={memoFilter}
-                  onChange={(event) => onMemoFilterChange(event.target.value)}
-                  placeholder="Поиск по описанию"
-                  className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                />
-              </label>
-            </div>
-
-            <div className="mt-4">
-              {transactionsStatus === 'loading' ||
-              transactionsStatus === 'idle' ? (
-                <StatusCard tone="neutral" message="Загружаем транзакции..." />
-              ) : null}
-
-              {transactionsStatus === 'error' ? (
-                <StatusCard
-                  tone="warning"
-                  message={
-                    transactionsErrorMessage ??
-                    'Не удалось загрузить транзакции.'
-                  }
-                  actionLabel="Повторить"
-                  onAction={onRetryTransactions}
-                />
-              ) : null}
-
-              {transactionsStatus === 'ready' &&
-              totalTransactionsCount === 0 ? (
-                <StatusCard
-                  tone="neutral"
-                  message="Для этого инструмента пока нет транзакций."
-                />
-              ) : null}
-
-              {transactionsStatus === 'ready' &&
-              totalTransactionsCount > 0 &&
-              filteredTransactions.length === 0 ? (
-                <StatusCard
-                  tone="neutral"
-                  message="Нет транзакций, подходящих под фильтры."
-                />
-              ) : null}
-
-              {transactionsStatus === 'ready' &&
-              filteredTransactions.length > 0 ? (
-                <ul className="space-y-2">
-                  {filteredTransactions.map((transaction) => {
-                    const isEditing =
-                      activeEditingTransactionId === transaction.id
-                    const hasDeleteError =
-                      deletingTransactionId === transaction.id &&
-                      deleteTransactionErrorMessage !== null
-                    const isDeletingThisTransaction =
-                      deletingTransactionId === transaction.id &&
-                      deleteTransactionErrorMessage === null
-
-                    return (
-                      <li
-                        key={transaction.id}
-                        className="rounded-lg border border-slate-200 bg-white px-3 py-3"
-                      >
-                        {isEditing ? (
-                          <form
-                            onSubmit={(event) => {
-                              void handleUpdateTransactionSubmit(event)
-                            }}
-                          >
-                            <p className="text-xs font-medium uppercase tracking-wide text-slate-500">
-                              Редактировать транзакцию
-                            </p>
-
-                            <div className="mt-3 grid gap-2 sm:grid-cols-2">
-                              <label className="text-xs text-slate-600">
-                                Дата
-                                <input
-                                  type="date"
-                                  value={editingTransactionDate}
-                                  onChange={(event) =>
-                                    setEditingTransactionDate(
-                                      event.target.value,
-                                    )
-                                  }
-                                  className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                                />
-                              </label>
-
-                              <label className="text-xs text-slate-600">
-                                Направление
-                                <select
-                                  value={editingTransactionDirection}
-                                  onChange={(event) =>
-                                    setEditingTransactionDirection(
-                                      event.target
-                                        .value as TransactionDirection,
-                                    )
-                                  }
-                                  className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                                >
-                                  <option value="OUTFLOW">
-                                    {toDirectionSelectLabel('OUTFLOW')}
-                                  </option>
-                                  <option value="INFLOW">
-                                    {toDirectionSelectLabel('INFLOW')}
-                                  </option>
-                                </select>
-                              </label>
-                            </div>
-
-                            <div className="mt-2 grid gap-2 sm:grid-cols-[minmax(0,0.6fr)_minmax(0,1fr)]">
-                              <label className="text-xs text-slate-600">
-                                Сумма
-                                <input
-                                  type="text"
-                                  inputMode="decimal"
-                                  value={editingTransactionAmount}
-                                  onChange={(event) =>
-                                    setEditingTransactionAmount(
-                                      formatMoneyInput(event.target.value),
-                                    )
-                                  }
-                                  placeholder="0,00"
-                                  className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                                />
-                              </label>
-
-                              <label className="text-xs text-slate-600">
-                                Описание
-                                <input
-                                  type="text"
-                                  value={editingTransactionMemo}
-                                  onChange={(event) =>
-                                    setEditingTransactionMemo(
-                                      event.target.value,
-                                    )
-                                  }
-                                  placeholder="Без описания"
-                                  className="mt-1 block w-full rounded-md border border-slate-300 bg-white px-2 py-2 text-sm text-slate-800"
-                                />
-                              </label>
-                            </div>
-
-                            {!isEditingTransactionAmountValid &&
-                            editingTransactionAmount.trim().length > 0 ? (
-                              <p className="mt-2 text-xs text-amber-700">
-                                Сумма должна быть положительной и содержать не
-                                более 2 знаков после запятой.
-                              </p>
-                            ) : null}
-
-                            {updateTransactionStatus === 'error' &&
-                            updateTransactionErrorMessage ? (
-                              <p className="mt-2 text-xs text-amber-700">
-                                {updateTransactionErrorMessage}
-                              </p>
-                            ) : null}
-
-                            <div className="mt-3 flex flex-wrap gap-2">
-                              <button
-                                type="submit"
-                                disabled={!canUpdateTransaction}
-                                className="rounded-md border border-slate-300 bg-slate-100 px-3 py-1.5 text-xs font-medium text-slate-800 disabled:cursor-not-allowed disabled:opacity-50"
-                              >
-                                {updateTransactionStatus === 'submitting'
-                                  ? 'Сохраняем...'
-                                  : 'Сохранить'}
-                              </button>
-                              <button
-                                type="button"
-                                onClick={resetEditingTransaction}
-                                className="rounded-md border border-slate-200 px-3 py-1.5 text-xs font-medium text-slate-600"
-                              >
-                                Отмена
-                              </button>
-                            </div>
-                          </form>
-                        ) : (
-                          <div className="flex items-center justify-between gap-3">
-                            <div>
-                              <p className="text-sm font-medium text-slate-900">
-                                {toTransactionMemoDisplay(transaction.memo)}
-                              </p>
-                              <p className="mt-1 text-xs text-slate-500">
-                                {formatIsoDateRu(transaction.occurredOn)} ·{' '}
-                                {toDirectionLabel(transaction.direction)}
-                              </p>
-                            </div>
-                            <div className="text-right">
-                              <p className="text-xs uppercase tracking-wide text-slate-500">
-                                {transaction.currency}
-                              </p>
-                              <p
-                                className={`mt-1 text-sm font-semibold ${
-                                  transaction.direction === 'OUTFLOW'
-                                    ? 'text-rose-700'
-                                    : 'text-emerald-700'
-                                }`}
-                              >
-                                {formatSignedAmount(
-                                  transaction.amount,
-                                  transaction.direction,
-                                )}
-                              </p>
-                              <div className="mt-2 flex flex-wrap justify-end gap-2">
-                                <button
-                                  type="button"
-                                  onClick={() =>
-                                    handleStartEditingTransaction(transaction)
-                                  }
-                                  disabled={isDeleteTransactionSubmitting}
-                                  className="rounded-md border border-slate-200 px-2.5 py-1 text-xs font-medium text-slate-600 disabled:cursor-not-allowed disabled:opacity-50"
-                                >
-                                  Редактировать
-                                </button>
-                                <button
-                                  type="button"
-                                  onClick={() => {
-                                    void handleDeleteTransactionClick(
-                                      transaction,
-                                    )
-                                  }}
-                                  disabled={isDeleteTransactionSubmitting}
-                                  className="rounded-md border border-rose-200 px-2.5 py-1 text-xs font-medium text-rose-700 disabled:cursor-not-allowed disabled:opacity-50"
-                                >
-                                  {isDeletingThisTransaction
-                                    ? 'Удаляем...'
-                                    : 'Удалить'}
-                                </button>
-                              </div>
-                              {hasDeleteError ? (
-                                <p className="mt-2 text-xs text-amber-700">
-                                  {deleteTransactionErrorMessage}
-                                </p>
-                              ) : null}
-                            </div>
-                          </div>
-                        )}
-                      </li>
-                    )
-                  })}
-                </ul>
-              ) : null}
-            </div>
-          </>
-        )}
-      </article>
-    </div>
-  )
-}
-
-interface MetricCardProps {
-  title: string
-  subtitle: string
-  totals: CurrencyTotalsDto
-}
-
-function MetricCard({ title, subtitle, totals }: MetricCardProps) {
-  const entries = Object.entries(totals).sort(([left], [right]) =>
-    left.localeCompare(right),
-  )
-
-  return (
-    <article className="rounded-xl border border-slate-200 bg-slate-50 p-4">
-      <h2 className="text-sm font-semibold text-slate-900">{title}</h2>
-      <p className="mt-1 text-xs leading-relaxed text-slate-500">{subtitle}</p>
-
-      {entries.length === 0 ? (
-        <p className="mt-6 text-sm text-slate-600">Пока нет данных.</p>
-      ) : (
-        <ul className="mt-6 space-y-2">
-          {entries.map(([currency, amount]) => (
-            <li
-              key={currency}
-              className="flex items-center justify-between gap-3 text-sm"
-            >
-              <span className="font-medium text-slate-700">{currency}</span>
-              <span className="font-semibold text-slate-900">
-                {formatAmount(amount)}
-              </span>
-            </li>
-          ))}
-        </ul>
-      )}
-    </article>
-  )
-}
-
-interface StatusCardProps {
-  tone: 'neutral' | 'warning'
-  message: string
-  actionLabel?: string
-  onAction?: () => void
-}
-
-function StatusCard({ tone, message, actionLabel, onAction }: StatusCardProps) {
-  const toneClasses =
-    tone === 'warning'
-      ? 'border-amber-200 bg-amber-50 text-amber-900'
-      : 'border-slate-200 bg-slate-50 text-slate-600'
-
-  return (
-    <div className={`rounded-xl border px-4 py-5 text-sm ${toneClasses}`}>
-      <p>{message}</p>
-      {actionLabel && onAction ? (
-        <button
-          type="button"
-          onClick={onAction}
-          className="mt-3 rounded-md border border-current px-3 py-1 text-xs font-medium"
-        >
-          {actionLabel}
-        </button>
-      ) : null}
-    </div>
-  )
-}
-
-function formatAmount(amount: string): string {
-  const isNegative = amount.startsWith('-')
-  const unsigned = isNegative ? amount.slice(1) : amount
-  const [integerRaw, fractionRaw = '00'] = unsigned.split('.')
-  const integerPart = integerRaw.replace(/\B(?=(\d{3})+(?!\d))/g, ' ')
-  const fractionPart = fractionRaw.padEnd(2, '0').slice(0, 2)
-  return `${isNegative ? '-' : ''}${integerPart},${fractionPart}`
-}
-
-function formatSignedAmount(
-  amount: string,
-  direction: 'INFLOW' | 'OUTFLOW',
-): string {
-  const prefix = direction === 'OUTFLOW' ? '-' : '+'
-  return `${prefix}${formatAmount(amount)}`
-}
-
-function isValidIsoDateValue(value: string): boolean {
-  return /^\d{4}-\d{2}-\d{2}$/.test(value)
-}
-
-function isValidPositiveAmountValue(value: string): boolean {
-  return /^(?:0|[1-9]\d*)(?:\.\d{1,2})?$/.test(value) && Number(value) > 0
-}
-
-function normalizeTransactionMemo(memo: string | null): string {
-  return memo ?? ''
-}
-
-function toTransactionMemoDisplay(memo: string | null): string {
-  const normalizedMemo = normalizeTransactionMemo(memo)
-  return normalizedMemo.length > 0 ? normalizedMemo : 'Без описания'
-}
-
-function todayIsoDate(): string {
-  const now = new Date()
-  const year = String(now.getFullYear())
-  const month = String(now.getMonth() + 1).padStart(2, '0')
-  const day = String(now.getDate()).padStart(2, '0')
-  return `${year}-${month}-${day}`
-}
-
-function formatIsoDateRu(isoDate: string): string {
-  if (!/^\d{4}-\d{2}-\d{2}$/.test(isoDate)) {
-    return isoDate
-  }
-
-  const [year, month, day] = isoDate.split('-')
-  return `${day}.${month}.${year}`
-}
-
-function toDirectionLabel(direction: TransactionDirection): string {
-  return toDirectionSelectLabel(direction)
-}
-
-function toDirectionSelectLabel(direction: TransactionDirection): string {
-  return direction === 'INFLOW' ? 'Доход' : 'Расход'
-}
-
-function toAccountTypeLabel(type: AccountType): string {
-  const labels: Record<AccountType, string> = {
-    CASH: 'Наличные',
-    DEPOSIT: 'Депозит',
-    FUND: 'Фонд',
-    IIS: 'ИИС',
-    BROKERAGE: 'Брокерский',
-  }
-  return labels[type]
-}
-
-function toAccountStatusLabel(status: AccountStatus): string {
-  if (status === 'ACTIVE') {
-    return 'Активный'
-  }
-  if (status === 'ARCHIVED') {
-    return 'Архивный'
-  }
-  return status
 }
 
 function normalizePersonalFinanceSnapshotCategoryOrder(
@@ -2392,41 +1074,6 @@ function resolvePersonalFinanceSettingsSelection(
   }
 
   return activeSnapshots[0]?.card.id ?? resolvedCard.id
-}
-
-function getAccountCurrencyOptions(): string[] {
-  const fallbackOptions = [
-    'USD',
-    'EUR',
-    'RUB',
-    'GBP',
-    'CHF',
-    'CNY',
-    'JPY',
-    'KZT',
-    'TRY',
-  ]
-  const intlWithSupportedValues = Intl as typeof Intl & {
-    supportedValuesOf?: (
-      key:
-        | 'calendar'
-        | 'collation'
-        | 'currency'
-        | 'numberingSystem'
-        | 'timeZone'
-        | 'unit',
-    ) => string[]
-  }
-  const dynamicOptions = intlWithSupportedValues.supportedValuesOf?.('currency')
-  if (!dynamicOptions || dynamicOptions.length === 0) {
-    return fallbackOptions
-  }
-
-  return [
-    ...new Set(
-      dynamicOptions.map((currencyCode) => currencyCode.toUpperCase()),
-    ),
-  ].sort((left, right) => left.localeCompare(right))
 }
 
 function readNavigationFromUrl(): NavigationState {
