+++ frontend/src/features/investments/index.ts
@@ -1,0 +1,30 @@
+export { InvestmentsView } from './InvestmentsView'
+export { InvestmentOverviewTab } from './InvestmentOverviewTab'
+export { InvestmentTransactionsTab } from './InvestmentTransactionsTab'
+export { InvestmentSettingsTab } from './InvestmentSettingsTab'
+export {
+  formatAmount,
+  formatIsoDateRu,
+  formatSignedAmount,
+  isValidIsoDateValue,
+  isValidPositiveAmountValue,
+  normalizeTransactionMemo,
+  toAccountStatusLabel,
+  toAccountTypeLabel,
+  toDirectionLabel,
+  toDirectionSelectLabel,
+  toInvestmentAccountCountLabel,
+  toTransactionMemoDisplay,
+  todayIsoDate,
+} from './formatting'
+export { INVESTMENT_TAB_COPY } from './model'
+export type {
+  AccountWithBalance,
+  CreateAccountStatus,
+  CreateTransactionStatus,
+  CsvImportStatus,
+  InvestmentTab,
+  LoadStatus,
+  TransactionDirectionFilter,
+} from './types'
+export { MetricCard, NestedTabButton, StatusCard } from './ui'
